requirements.txt
Flask==1.1.2
openai==0.2.6
python-dotenv==0.15.0
requests==2.25.1
flask-cors==3.0.10